Refactor repository layers and cleanup
