+++ TheBackend.Api/Controllers/GenericController.cs
@@ -1,6 +1,8 @@
-﻿using TheBackend.DynamicModels;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
+using TheBackend.Application.Repositories;
+using TheBackend.DynamicModels;
+using TheBackend.Infrastructure.Repositories;
 
 namespace TheBackend.Api.Controllers
 {
