--- TheBackend.Api/IGenericRepository.cs
@@ -1,54 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace TheBackend.Api
-{
-    public interface IGenericRepository<T> where T : class
-    {
-        Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetByIdAsync(object id);
-        Task AddAsync(T entity);
-        Task UpdateAsync(T entity);
-        Task DeleteAsync(object id);
-    }
-
-    public class GenericRepository<T> : IGenericRepository<T> where T : class
-    {
-        private readonly DbContext _dbContext;
-        private readonly DbSet<T> _dbSet;
-
-        public GenericRepository(DbContext dbContext)
-        {
-            _dbContext = dbContext;
-            _dbSet = _dbContext.Set<T>();
-        }
-
-        public async Task<IEnumerable<T>> GetAllAsync() => await _dbSet.ToListAsync();
-
-        public async Task<T> GetByIdAsync(object id)
-        {
-            return await _dbSet.FindAsync(id);  // Assumes single primary key; for composite, pass array of objects
-        }
-
-        public async Task AddAsync(T entity)
-        {
-            await _dbSet.AddAsync(entity);
-            await _dbContext.SaveChangesAsync();
-        }
-
-        public async Task UpdateAsync(T entity)
-        {
-            _dbSet.Update(entity);
-            await _dbContext.SaveChangesAsync();
-        }
-
-        public async Task DeleteAsync(object id)
-        {
-            var entity = await GetByIdAsync(id);
-            if (entity != null)
-            {
-                _dbSet.Remove(entity);
-                await _dbContext.SaveChangesAsync();
-            }
-        }
-    }
-}
