+++ TheBackend.Application/Repositories/IGenericRepository.cs
@@ -1,0 +1,11 @@
+
+namespace TheBackend.Application.Repositories;
+
+public interface IGenericRepository<T> where T : class
+{
+    Task<IEnumerable<T>> GetAllAsync();
+    Task<T> GetByIdAsync(object id);
+    Task AddAsync(T entity);
+    Task UpdateAsync(T entity);
+    Task DeleteAsync(object id);
+}
