+++ TheBackend.Infrastructure/Repositories/GenericRepository.cs
@@ -1,0 +1,45 @@
+using Microsoft.EntityFrameworkCore;
+using TheBackend.Application.Repositories;
+
+namespace TheBackend.Infrastructure.Repositories;
+
+public class GenericRepository<T> : IGenericRepository<T> where T : class
+{
+    private readonly DbContext _dbContext;
+    private readonly DbSet<T> _dbSet;
+
+    public GenericRepository(DbContext dbContext)
+    {
+        _dbContext = dbContext;
+        _dbSet = _dbContext.Set<T>();
+    }
+
+    public async Task<IEnumerable<T>> GetAllAsync() => await _dbSet.ToListAsync();
+
+    public async Task<T> GetByIdAsync(object id)
+    {
+        return await _dbSet.FindAsync(id);
+    }
+
+    public async Task AddAsync(T entity)
+    {
+        await _dbSet.AddAsync(entity);
+        await _dbContext.SaveChangesAsync();
+    }
+
+    public async Task UpdateAsync(T entity)
+    {
+        _dbSet.Update(entity);
+        await _dbContext.SaveChangesAsync();
+    }
+
+    public async Task DeleteAsync(object id)
+    {
+        var entity = await GetByIdAsync(id);
+        if (entity != null)
+        {
+            _dbSet.Remove(entity);
+            await _dbContext.SaveChangesAsync();
+        }
+    }
+}
